Finished the general look of the main screen
Still need to flesh out some ideas regarding privileges. To look at the rest of the app just change the line in MainActivity.kt back to what it was. It is going to currently only display Select Service Screen! Happy with the other images, but I might change the in progress/under construction one.