Added the info box and made some other stuff work
Still need to try to get the desired effect for the InfoBox/Login Screen. I also need to add a back button for the login screen!

There should be a way to log out.  I believe this is already implemented if I recall correctly.

NOTE: This is still not epic, and lots of code here is volatile as I am not sure what path to take for managing some ideas.